Removed controllers
